DAQ-470 Remove Sub-Server dependencies on LocalProperties
The Object, Log and Channel server currently each trigger the loading
of the LocalProperties files causing three filesystem operations instead
of one and allowing the potential for the properties to be different
from server to server.

Local Properties loading has now been moved to before any of these three
servers are started and the onlt the relevant properties are now passed
in as arguments to main() or System properties. This allows the following
changes:

ChannelServer:
 Remove call to setLogDirectory which triggered loading Local Properties
 Convert LocalProperties refs to System Properties ones
 Remove translation of Corba keys.
 Allow threadPoolSize to be passed in to GdaEventChannelImpl
 Add some logging

LogServer
 Remove call to setLogDirectory which triggered loading Local Properties
 Convert LocalProperties refs to System Properties ones

ObjectServer
 Remove call to setLogDirectory which triggered loading Local Properties
 
Change-Id: Id9bf0c6d2c8e47fd71523817c8d2fc089e88b60c