I05-166 Move things which don't need the UI thread out of it
Before this change every update received from the the server would
require the UI thread to be processed even when it didn't update the GUI
e.g. In the case of baton renew requests.

This change refactors the code so only calls which actually need the UI
thread are made in it. This should significantly improve the performance
of the GUI, especially under high event rates.

Also:
- Simplify the setStatusLineText method removing the varargs and
  removing the tooltip option which is never used.
- Remove unneeded cast to IActionCommandMappingService

Change-Id: If98372f26f9c61082e52bedd50bf261314c4e084