I05-166 Only update the GUI with ScanEvents at a reasonable rate
When running fast scans (>~20 Hz) the rate of scan events received can
exceed the rate at which the GUI can update. This is mostly because each
update received is handled in many places, many of which need the UI
thread.

This change limits the rate of attempted GUI updates for scan events to
20Hz. This will mean that some updates are dropped however during a fast
scan seeing every single point in the status bar is not useful, and
waiting for a long time after the scan has finished for the UI to catch
up is bad.

Also:
- Rearrange if statement to share repeated logic

Change-Id: Ib67dd6d2072bf139a51c1777a9ccb16676bfca9b