DAQ-515 Remove unnecessary configured flag from EventService
As the configured flag is always set true at the end of the private
constructor. The only way for it not to be configured is if an exception
is thrown during the constructor. If this happens the exception will be
rethrown wrapped as a RuntimeException by the getInstance method.

So the getInstance, isConfigured pattern used elsewhere in the code was
redundant. If getInstance works then isConfigured will always be true.
These are also removed here.

Change-Id: Ic8302311a763ea7aa3f6b67d1487379a836b31ba