Don't fail server product export if tp/artifacts.xml missing
The server product export copies tp/artifacts.xml, but this file is copied
as a record only, and is not actually required for running the server.

When developers make changes to tp/ in their local workspace, this sometimes
involves deleting the tp/artifacts.xml file. Normally, that is recreated by
Eclipse, but if it is not, the product export should not fail.

Previously, if the file was missing, the product export would fail with
the error "Cannot perform operation from directory to file".

Change-Id: Iba42c714421d9597101ea3db17fe5bc6a39db969