Add control points to DeferredScannableGroup for validation
Adds optional ControlPoints for the purposes of "verifying" a
deferred coordinated move with the controller.

It's really just a set of horrible hacks to workaround problems
with the controller on I16 (axis vibration and move requests
within the deadband can result in a "stuck" move).

There's quite a deep inheritance chain that this the
DeferredScannableGroup class is part of that prevents nice
encapsulation of this change. It'd be nice to reorganise things.

Change-Id: I93141ec8f697f3a6d837f53e5c490671917905f7