I15_1-367 (9.4pre) Fix 'h' to re-histogram key binding in LiveStreamView
Rather than configuring the plotting system with no actionBar, preventing
any key bindings from being made, start with the standard plotting system
action bar and remove all of the ToolBar and Menu contributions we do not
want to see in this view.

By defining the items we want to keep, rather than those we want to
exclude, this change should be robust to new contributions we are not
interested in.

Change-Id: I0682ed391301c3a0074afab66eea8aa4b25c31d3