DAQ-596 Add benchmarks for Corba and JMS events
There are 2 benchmark classes both using the EventService interface for
benchmarking Corba and JMS events. There are 2 benchmarks for each. The
endToEnd benchmark test sending and receiving a message to test total
throughput. The sendEvents benchmark tests only how fast can you publish
events, it is relevant for code like the scan which needs to dispatch
events and shouldn't be blocked. To run the benchmark you need a
ChannelServer running for Corba and an ActiveMq server running for JMS

Note: A current deficency of this benchmark is the performance of the
corba "batchMode" is not tested in the end to end mode so the
performance seen with corba might be lower than the real performance. I
can't think of a way to test this. Maybe using JMH multithreading test
somehow?

Note 2: There are still some issues closing the JMS events benchmarks I
think this is due to some ActiveMQ thread pool not being closed. Suggest
a shutdown method could be added to the EventService to allow this to be
fixed.

Change-Id: If92d5fe384e331477a59d065f72ad159fe34461d