DAQ-613 Add queue to JmsEventDispatcher
This ensures callers to publish will not be blocked. The actual
serialisation and dispatching of messages over ActiveMQ is moved into a
separate thread managed by an ExecutorService.

Also adds null check before checking if Serializable to allow null
messages to be sent.

If the dispatch thread dies a new one will be started in its place.
Hopefully this is a rare case. There is a weird logging bug in this code
commented with the behaviour.

Change-Id: Ie9d80aa92736184b91787572ca84d4f366f2e864