B18-126 : Move data collection for 'Detector Rates View' to server.
A new server side class DetectorMonitorDataProvider has been created to 
provide the data used in the 'Detector Rates' view; the data collection in XspressMonitorView and XMapMonitorView classes is now carried out using 
the new class. This refactoring enables explcit control over 
data collection for the rate view from the server.
The thread handling in MonitorViewBase has also been simplified - the collection
thread is started when the 'start' button is pressed and ends when 'stop' button
is pressed, the data provider prevents collection, or an exception is thrown.
A 'set collection time' button has also been added to the rate view gui.

Change-Id: I9c03135ebe54b766137d44cf1b9b89be700db1e5