I05-166 Trying to improve JythonTerminalView performance.
Removes UpdateRunner and replaces it by an scheduled execution of a
SimpleOutputUpdater at 20Hz. This removes the need for the inqueue flag
in the SimpleOutputUpdater.

The SimpleOutputUpdater is modified to reduce the amount of time spent
in the UI thread. Now needs to use syncExec not async to ensure the
update of the text happens before the scrolling and moving to the top.

Change-Id: I23c104d2dc7ef825dd1cc911043fb0fe1465b7f9