DAQ-280 Extract the JPA persistence classes from core into a new plugin
The JPA persistence classes used with Derby and Toplink need to be
included in a fragment for them to be visible in OSGi. Create a new
plugin for this purpose that only contains these classes to ensure that
the (yet to be produced) fragment only has to depend on relevant code.

At the moment this plugin contains a copy of the gda.util.persistence
classes within differently named packages so that the many classes and
scripts that refer to the originals are still valid. Migration to use of
the new plugin's versions of the classes will be carried out piecemeal
via subsequent changes. A corresponding test fragment is also created.

Change-Id: I878543e110b143ae1d37781ee90da46b7cfb4185