DAQ-604 Reorganise persistence plugin to contain all relevant code
In order to support the switch to using uk.ac.diamond.daq.persistence as
the manager for toplink and direct derby database access and correct
operation from all clients, the following changes have been made:

1. Include the BeamlinConfigurationManager classes
2. Re-arrange the packages so that there is a separate sub-tree for
each of the Persistence units defined in persistence.xml
3. Include the toplink jars to simplify class resolution
4. Update persistence.xml and tests to reflect the new package trees
5. Modify LocalPersistence.java to ensure the correct Classloader is
always used to create the JPA Entity Manager
6. Update the docs with the new pacakge paths

This plugin currently contains copies of classes in uk.ac.gda.libs and
uk.ac.gda.coreplus the appropriate xml files so that those classes are
still available to the build maintaining their functionality. These
will be removed in a future change when the switch over to using the new
persistence plugin is submitted.

Change-Id: I17b5581392ac8ad10cc428cf07d613f2de86b19a