DAQ-604 Add Derby fragment to allow loading of persistence classes
Because the Derby database needs access to the classes in
uk.ac.diamond.daq.persistence, the classpath of the Derby bundle needs
to be extended by adding a fragment that depends on the persistence
plugin.

To ensure it is present in server materialisations when required, add
details of the fragment to the uk.ac.diamond.daq.persistence cspex file.


Change-Id: Ibe74809412265536d031ef5dbfaafcb1e3867b98