I15_1 (9.5pre) Add the outputEveryArray option to AutoSummingProcessDecorator
Normally AutoSummingProcessDecorator will configure the Proc plugin to
only generate a single array at the end of the summing process.

The outputEveryArray option causes all of the intermediate, partially
summed, arrays to be output too. This can be used in conjunction with
the Area Detector POS plugin and SWMR (single writer, multiple reader)
HDF5 plugin to write the intermediate images into the same location in
the hdf5 file, leaving only final, fully summed, images being left in
the file at the end of the scan.

For images which take minutes or hours to acquire, access to the
intermediate image data can allow scientists to detect poor samples or
failed collections early, and avoid loss of beam time.

Change-Id: I2ab6256b857a9ed4ca2b7d6b501425f4e05fc61b