DAQ-685 Remove duplicated EPICS jars
In order to achieve this the Live Streams view has been moved into a
new plugin. This new plugin has an optional dependency
on uk.ac.gda.epics where the EpicsV3DynamicDatasetConnector has been
moved to. This is to allow the MJPEG stream feature to be used in 
clients which don't want to depend on EPICS, however if the EPICS plugin
is available in the product then the EPICS stream can be used in 
addition.

Add detection and logging if EPICS array PV are defined but the EPICS
bundle is not available.

Adds the new plugin to the uk.ac.gda.client.feature and adds
uk.ac.gda.epics to the p45 client to allow EPICS streams to be used.

Change-Id: Ifc92c2e686e39d8c54f61cf538ea51d0cdd13ea6