Fix bug where server startup output was not displayed in the client console
This bug was introduced by a performance improvement which removed
locking. The locking enforced this would work correctly.

This will now work reliably but the behaviour is subtly changed from
before. Previously the startup console output would disappear when the
first output was received, now it will just scroll off with the other
output, which I prefer anyway.

Change-Id: I3bf0e4ed2afdc2e62b017ca777c90694ad041b45