MXGDA-3023 RotationViewer: really don't popup when user reclicks move button too soon
This improves on and fixes change 6021 (itself a fix for change 5536),
which didn't work because the exception with the "was already busy" message
could be wrapped, particularly by ScannableBase.moveTo:

  } catch (Exception e) {
    // convert to a device exception
    throw new DeviceException("Move failed for " + getName() + ".moveTo("+position+")", e);
  }

First improvment is to recurse through the causes, logging without a popup
if already busy, but with a popup for some other error, such as a limit
violation.

Second improvement is to report the original cause of the exception,
which is often more detailed than ScannableBase.moveTo's "Move failed".

Third improvement is to always log the whole stacktrace at DEBUG, then
tersely log the exception at info (already busy) and error (not already busy).

Fourth and final improvement is to use parameterised logging.

Change-Id: I5666598b7f05d77bc595b3d02ca3422a33f02ad0