Use Set and Collection for defaultScannables instead of Vector
The interface doesn't need to specify anything more than collection and
there is no need for a fully synchronised implementation to store the
scannables. CopyOnWriteArraySet ensures that modifications are thread
safe without the overhead of synchronisation while reading.

get_defaults should return a copy of the list so that defaults can only
be modified by the server.

Change-Id: I88407e29ac2db0052be925516bf414de1db97943