Add default_scannables context manager
For when series of scans in a script all require the same default
scannables. Saves needing to add and remove scannables manually.

with default_scannables(temp1, temp2):
    scan x 0 10 1
    scan y 10 0 1
    rscan z -5 5 1

If a scannable is set as a default before the with block, it is left as
one afterwards. eg

add_default(temp)
with default_scannables(temp):
    scan x 0 10 1

is the same as just running the scan.

Change-Id: Iea763d95dbfa289d92b4d2c1cd1ae4bbf168aa61