[DAQ-758] Correct calculation of device timeout
Modify DeviceRunner.getTimeout() so that, if no explicit timeout
is set for a detector, it is set to (exposure time + 1s), rounded up
(Math.ceil()), rather than (potentially rounded down (Math.round()).

Also remove the test for IDetectorModel, as this is an extension
of ITimeoutable, which is tested for above, so this code will never
be executed.

Signed-off-by: Anthony Hull <anthony.hull@diamond.ac.uk>