DAQ-743 Don't set the var/cache dir in GDAJythonInterpreter from Spring
Currently ALL GDA9 beamlines have the two lines in their Jython setup:

<property name="gdaVarDirectory" value="${gda.var}" />
<property name="cacheDirectory" value="${gda.var}/jythonCache" />

which (via JythonServer) set the _gdaVarDir and cacheDir fields on the
GDAJythonInterpreter class. This is completely redundant as the static
initialiser block does this anyway for the cacheDir and the getGdaVarDir
method does it on the fly for _gdaVarDir.

This complicates the Jython spring setup unnecessarily and is clearly
not being used for its intended purpose as everyone uses the default
value.

There are 4 older beamlines that replace the {gda.var} with {gda.config}
/var but 1 of these has just been switched off, another is being rebuilt
from the ground up and the other two only have this setting for
historical reasons and Iain is happy to change this.

This change turns off the setting of the fields by the JythonServer
class which will allow the property lines to be removed from the Spring
Jython setup file for all beamlines and will also allow the construction
and configuration of the Jython Server and Interpreter to be simplified

Change-Id: I00fdee275160ac0e2cbbfee1fdcf84e6159be2f7