[DAQ-707] Refactor DummyXmapController
Done in order to be able to create an alternative version that
uses an Area Detector to write data. This allows a better simulation
of the situation on I08, where the XMAP detector transmits data
to an Area Detector running on another machine.

This change also ensures that DummyXmapController(Base) extends
the same base class (DeviceBase) as the Epics version.

Change-Id: I8009036bb3c5a258d1915781c6e3a25292e37e11