DAQ-806 - Add SSH access to server
If a key directory is given (via property) it is used to look up a
public key. For a user abc12345 logging in to beamline iXX, the public
key is expected to be in either of:
    ${gda.remote.ssh.keys}/abc12345.key
    ${gda.remote.ssh.keys}/iXX/abc12345.key

If the directory property is not set, no authentication attempt is made
and all connections are accepted.

Change-Id: Ib8b1813fcc384eeb2492b5e879ea57d8099cd1aa