I20-224 : Improvements to Fluorescense detector view
 - Bug fix for invalid thread access error that prevents MCA data from
 being saved if using the 'Save on acquire' option. If using 'Save on acquire',
the new file(s) are now saved to the default location each time; if
'Save on acquire' is *not* being used, the save file dialog is opened (as before).

- The detector output options part of the gui is now hidden according to the local property, 'exafs.outputeditor.showFluoOptions.preference' (same as original detector
view). i.e. if set to true, this indicate that another gui is used to set output
options, and they are hidden in the Fluorescence detector view.

Change-Id: I858718dfed5759361e7764983c2148819cf228a4