I20-225 : Detector rate view ionchamber values should not be dark current corrected
This changes ensures that the ionchamber count values displayed are not
dark current corrected. This is important o avoid misleading values being
displayed when using the detector rate view to monitor counts during
ionchamber amplifier adjustment.
For each rate view collection, the latest state of the dark current
on/off option is stored before switching it off, collecting the data and
restoring it to its original value at the end.

Change-Id: I977af88d79b4e38b585b8e9d44148b92834db018