I20-201 : Add realistic unit tests for QExafs, Xas scans
Unit tests have been added to test Xas and QExafs scans. These are setup
using XML files, the same as is typically done on the beamline using the gui.
The detector data in the Nexus file is checked to make sure the correct
datasets are present and have correct dimensions for the detector configuration(XSpress2, Tfg/Ionchamber related detectors only at present).

Change-Id: I6b66c049ecaca318c0fb94219fc1678933f36c5b