[I08-57] XMAP runnable device to use Area Detector to write files
The XMAP detector on I08, for which this runnable device is written,
is in three parts:

- The XMAP controller itself (Windows)
- A transmit (TX) plugin (Windows) which acts as a bridge for
transmitting data to...
- ...a receive (RX) plugin (Linux), which implements a restricted
form of Area Detector (pos & hdf5 modules)

The current XMAP controller does not write data in the format
required for processing, but the Area Detector's HDF5 plugin
will do so. This runnable device therefore still uses the XMAP
controller to control the device, but the Area Detector to
write the data.

Change-Id: Iad73dbb0c49890e8994953e916c67b410ae8050a