B18-147 : Initial working version of spreadsheet view
This is an initial commit of the spreadsheet view for B18.
Scans are specified by the usual series of 4 xml files, and 
parameters from xml file can be selected for alteration. Each selected
parameter and its new value are shown in the spreadsheet view. New xml
files can be generated that have the user specified values for the given
parameters.
Currently, setting new value for a parameter works correctly when values
are single number or string. Still to do: how to get user input for parameter
that expects a list (i.e. a list of sample stages to be moved).

Change-Id: I5a012cbd0981a845b8d32b8ee9edabdba96ceba6