DAQ-621 DAQ-727 Fix problems introduced by child consoles
Changes made in d20f463 and d486b0d1 made commands specifying an input stream
run in their own GDAInteractiveConsole. This introduced problems in
maintaining state from one to another (especially where Jython accesses
some attributes statically eg PySystemState#setdefaultencoding).

The input stream to use is now stored in a ThreadLocal in the Console
allowing commands to be run in the same console but keep using their own
STDIN. If stdin is not set, the default InputCommands#requestInput is
used.

Change-Id: Ib6fb90aba3898c6e129307664cabe1f46d48e2b6