B07-163 Show error dialog if setting the sequence fails
When the sequence was invalid (e.g. duplicated region name) then an
exception was thrown however this was not displayed to the user so it
appeared to fail silently.

Now an error dialog is displayed with the message from the exception and
the sequence that caused it.

Added doc to the analyser interface

Change-Id: I061e7da76a20298c3af2e4fcaff606e12bd0bff1