DAQ-825 Include outer scannables when showing number of scan points
StatusPanel now gives total number of scan points as the size of the
compound points generator (the same method used in ExecuteView), which
includes outer scannables. PathInfo still used to get smallest steps in
path, and to get 2D map scan points (as before) should a GeneratorException
be thrown.

updateStatusLabel() is now additionally called whenever
the outer scans are changed (as long as they are included in scan) i.e:
 - When an outer scannable is included/excluded via checkbox
 - When model is edited through the MultiStepEditorDialog
 - When model is edited in the text widget

Also:
 - Passing mapping bean to StatusPanel, not the whole view
 - Moved MappingExperimentView::getPointExposureTime to StatusPanel

Change-Id: I2ae44ef7c45913d0d2b8363a0d12ff592e8c1126