Add logging to ScannableCommands to allow understanding of usage
This adds debug level logging to the calls of public methods in the
ScannableCommands class. These methods are mostly invoked via user
interaction via Jython. In order to cleanup this class and improve it in
the future we need to understand how it's used. Over time this logging
should help to reveal this.

Change-Id: I17772201afe50d8f56bd37f9afc946160540d865