B18-175 : Check number of detector elements in XML when adding scan to queue
A consistency check is now carried out when adding scan from experiment
explorer to the queue. If the detector specified in the xml file cannot be
found, or if the number of elements in the xml is different from the number
on the detector, then a warning with some (hopefully) useful information
is displayed.

Change-Id: I5c3dea85dd050877e0ad717c4c42ec8a24ecfd15