change ScannableStatus to be a simple enum
Currently ScannableStatus has a status and a scannable name. The status
is an int acting as an enum, where the value should be one of three int
defined in the class (but of course, this can't be enforced with ints).

The scannable name is set but there is only one place in the whole
codebase of GDA where this is used, namely PVScannableBean.update()
in the swing client (gda-legacy repo). It uses the name as a sanity
check, in fact the check should always pass - see the change for this
topic in that repo.

Change-Id: I24fc5d43b28dd54a4976c7f7d8151df56b429f75