DAQ-836 Add implementation of PropertyService
This also requires a change to ConfigurationDefaults to ensure the
relevant system properties are set before attempting to load the
property files. Achieved by converting the initialiseServerEnvironment
method to a static initialised run when the ConfigurationDefaults enum
is first accessed.

Switches on support for DS annotations to auto-generate component.xml
files

Change-Id: I7e84cea0c0595d590988bd4c24d52dc24a322de1