I20_1-150 : Make output of TfgScalerWithLogValues more configurable
Output of TfgScalerWithLogValues is now configurable by use of a new
class ScalerOutputProcessor, which is used to replace the readout method
and to set the values of the 'extraNames' and 'outputFormat' arrays. 
This allows the readout of the scaler to produce a set of single scaler
values, ratios of scalers, or log ratios in arbitrary orders and combinations.
Unit tests have been added to test the new functionality.

Change-Id: I81a1e2ada9c063808339236dd6934d5dbda06f74