B18-179 : Fluorescence detector view : add transition line for selected element
New gui controls have been added to Fluorescence detector view to select
an element and line to be shown in the plot; the energy of the line is also
displayed, and a checkbox can be used to show/hide the line in the plot.
If there is a scan currently selected in the 'Experiment explorer', the
edge settings will be taken from the scan parameters (i.e. element name
and edge from the QExafs, Xas, Xanes xml file) and the line added to the
plot by default. Xraylib is used to provide the atomic data.

Change-Id: I80c2bdbace6bbd07a0fc747408456e1acfd92a92