I20_1-37 : Update Xspress3 to use latest Epics version
The Xspress3 PV provider has been updated to include the PVs used
by the latest release of the Epics XSpress3 module. In particular, the
names of the PVs used to return time series 'SCA' data from the various
areadetector plugins have been updated.
These new PVs can be used by setting 'useNewEpicsInterface' to true when
creating the controller.
Other changes : Link to hdf file in the Nexus file is now set correctly,
scaler data is added to file with the correct shape, number of frames
to collect is also set in the hdf file writer plugin. 

Change-Id: I205a73f10777d902938a1310301ccd5e110bbac9