Fix problem where one selected job determines the states of all buttons
Instead of setting the state of every button for every job in the
selection, set each button once according to the overall state of
the selection.

This makes it clear when buttons are enabled and why. It also makes it
obvious when too many functions are overloaded onto a single button.

To facilitate this, make getSelection() return a list rather than an array.

Also, update the actions when the viewer input is updated to ensure that
the actions are correct even if the the selection hasn't changed.

Change-Id: Ib79a344ddbc554b59eb88ab1168a71471b397449