[DAQ-1019] Make XmapMappingModeEpicsLayerImpl more robust to config failures
The constructor of XmapMappingModeEpicsLayerImpl previously did
a lot of work, some of which could throw an exception, which in turn
prevented GDA from starting up.

This was noticed on I08 when XMAP was switched off.

This change moves most of the code from the constructor
into configure(), where exceptions do not stop GDA starting (if the
appropriate property is set).

Also delete unused collectMode

Change-Id: I12752486826c41eca427577bba60023383a0c31b