DAQ-941 Also handle aliases when using SpringApplicationContextBasedObjectFactory
This changes the way the factory works to ensure aliases are also
handled. The new approach is to construct a Map of all the Findables
and aliases known to Spring in the constructor of the factory. This is
then always used internally when finding Spring objects.

This ensures that the getFindableNames method and getFindable are
consistent i.e. Only passing names returned from getFindables to
getFindable will return an object.

Change-Id: I7053270011c5ce18beeef7e88e48a1a46d6a3ce5