DAQ-1041 Add load/save support for Beamline Configuration
Adds support to serialise beamline configuration to json, saving to a file
with extension *.pos in default directory IFilePathService.getPersistenceDir()
+ "/blconfigs" (created if not there)

Note that serialisation is to/from Map<String, Object> - not ControlTree,
which would ignore stored positions.

Buttons now integrated into the ControlTreeViewer toolbar (but they only
appear in *this* instance of the widget).

Will now warn users if they're trying to overwrite a configuration

Change-Id: I2386fcebe9c63304b6f9c0b6de4a663cce8bb55f