DAQ-1075 Get axes names for region/path composite labels in Mapping UI
This change does a couple of things in the Mapping UI:
1) Move the Mapping Stage Manager from a separate tab to a dialog. The
dialog is opened through a button (another cog) next to the Scan path
combo box. The dialog does not show Beam Size - this is still part of 
the MappingStageInfo (why?) but does not need to be modified once con-
figured in Spring.

2) This new dialog looks for a MappingStageOptions Spring configuration
specifying allowed motors for fast, slow and associated axis. If the
configuration is there, these can be selected from combos. Otherwise,
the dialog draws text widgets like before. Warning: there is NO verifi-
cation in the latter version - Combos are prefered.

3) The labels in the region and scan composites now pick up the names
of the motors set in the MappingStageInfo. When you press OK in the
dialog, these composites are redrawn to pick up the changed axes.

 - very minor and not worth enumerating: removed colons from labels for
 consistency.
 
 - Update: If MappingStageOptions has been configured without associated-
 Axes, the label and combo widgets associated with this are not drawn.
 
 - Update: Handles the case of axes changing when selecting a Malcolm
 device.

Change-Id: I865f5a0adf2ede33f24197d159dcbf6d353e8799