I15_1-444 (9.7pre) Fix unexpected mscan step behaviour
Scans using the step generator should always use the given step value,
even if the step value is inconsistent with the given stop value.

Add a test case with start/stop/step values which are inconsistent, so
behaviour under these circumstances is both defined and meets user
expectations.

Change the behaviour to pass these tests. For example, previously
step(... start=0, stop=5, step=2) would result in a sequence of 0, 2.5, 5
but now it will generate a sequence of 1, 2, 4 as expected.

Also to confirm that RasterModel is behaving as expected, add a few more
test points to RasterTest.testFillingAMoreComplicatedBoundingRectangle().

Change-Id: I4b9ee6c1e27ae2901a4192dd40c58af0dcf7d0d1