DAQ-588 Make client feature depend on common feature
The feature uk.ac.diamond.daq.common.feature is intended to contain
things common to the client and server, however before this change the
client did not actually include it.

Now it depends on the common feature remove direct dependencies on
Apache commons plugins.

This is a prerequisite to moving ActiveMQ into the TP

Change-Id: I35aa0c76061c3dd4a8e8a5f060c1567f95f0019c