Performance improvement when calling SpecsPhoibosAnalyser.getImage
This was prompted by looking at a problem reported by i09-2 where a
scan failed. The error logged was about the mismatch of the image array
length. This change will avoid that and will also improve performance as
we only ask for the energy channels and slices once from EPICS instead
of twice.

Change-Id: I4b1eb3aae0d2618a807e1ed2ef3a900ceec2507d