[DAQ-1065] Change error handling in ScannableMotor
The previous behaviour depended on the setting of the property
gda.device.scannable.ScannableMotor.isBusyThrowsExceptionWhenMotorGoesIntoFault

- If true, isBusy() and waitWhileBusy() would throw an exception if
the underlying motor was in a fault or limit violation state.
- If false, neither of these functions would throw an exception.

This change removes support for the property such that:

- waitWhileBusy() always throws an exception if the motor goes into
fault/limit state
- isBusy() throws an exception only if it cannot get the motor state

This allows you to see the status of a motor, and potentially move it
away from a limit violation, but also to know (because an exception
will be thrown) if a moveTo() (i.e. pos command) has failed.

Change-Id: I0e99781883f35cb40c4d15d1844c0372a8d6a268