I20-302 : Filename updates for XSpress4 detector
Path to output directory used for Hdf file writer is now set at start of scan;
the 'nexus' directory in top level of the visit/commissioning folder is
used as default if no value has been set.
The link to the the Hdf file added into the Nexus file at end of the scan is
now uses a relative rather than absolute path.
A new function ('getLiveStats') has been added - this returns an array of
data suitable for use in MontitorViewBase to view live count rates.

Change-Id: I6620143d2b66fd996290f3fea8787504f554ec5b