DAQ-1139 Adds server available watchdog to the client
This checks that the server can be reached every second. If the server
can't be reached then a dialog is shown to the user informing them. It
allows the client to be closed. The user can dismiss the dialog and
return to the client however, mostly likely it will not work.

Also adds an simpler improved dialog if the server can't be reached at
client startup.

Change-Id: I638341a3ff454b06e83a91ccf2e9ff28bbbd34f8