I20-318 : Updates after XES testing with Medipix
- Improve method used in ADRoiCountsI0 for getting ROI counts from detector
data. This helps account for the fact that the position of the ROI counts
value in the detector data can change depending on which stats are enabled
in the ROI stats plugin, and also which other plugins are being used.

- In collection strategy, set the ionchambers frame time array using the correct
time per step. This avoids a problem where the getCollectionTime() method of
ionchambers sometimes returns zero and so cannot be relied on to provide
the 'time per step' for the scan.

Change-Id: I0f8f480e1397696c750b1a377978275f2c6044c0