Finder improvements
Change to use a CopyOnWriteArraySet to hold the factories to avoid
copying the list for each find. As the factories are typically set at
startup and then the finder is used this should provide better
performance.

Adds logging when the factories set is modified and when an object is
found indicates which factory provided it.

Change-Id: I50bde568e54362d2dfc09642f0f5b22e873e0f9f