[DAQ-1102] remove reflective comparator for getQueue
remove ability to create queue with comparator. Clients can easily
sort the list on the field they want afterwards - turns out that
this code is not used, so the awful reflective comparator can be
removed. The only field called was submissionTime, by getStatusSet
which can do its own sort of the list after getting it.

Change-Id: I7ce49fb0ac4b5507567ba261061007debbc7f4f4