Move scan control button logic out of JythonTerminalView
The JTV was just a listener that passed events on to the ControlFactory.
Moving the logic into the Factory itself allows it to better reflect the
state of the server and prevent pause scan button being checked when no
script/scan is paused/running (or being unchecked if it is).

Change-Id: Ief2053756eaab16778ddf88be10105c25a6c2e7b