Check if script is paused while running scan
When pauseCurrentScript is called, the script only pauses when an explicit check
is made in the script. This checks the script pause flag at the same time
as the scan pause flag when a scan is started from a script.

If a scan is run manually while a script is paused it is not affected.

Change-Id: I8e2baca58aad0f0ebf4bbd5f4c32bab81dc5e997