DAQ-1170 Create RmiProxyFactory
This is the RMI equivalent of the Corba AdaptorFactory. It can provide
RMI proxies to access remote objects. It can be used seamlessly by the
Finder to provide auto-exported objects.

Also adds an integration test that uses the RmiAutomatedExporter and the
RmiProxyFactory together. This also tests that events sent from the
server objects reach the client proxies. To enable this testing change
rmiPort to be non-static.

Change-Id: I3ee45bcd628ace522b13148316f34b97d05ef9c6