DAQ-1230 Add a bridge from Jython logging module to SLF4J
Using the built in logging module to log add log messages now redirects
messages to the main GDA logs. Anything logged at INFO or above is
also printed to the console.

For logging exceptions, the built in exc_info keyword in jython logging
will log the full stack trace to the main logs and print a list of
causes to the console.

Change-Id: I9b22c428c9fd182530105c2ad3fe1b405e5b524b