DAQ-1271 Ensure Jython bundle classloader is always the TCCL for Jython threads
The threads which actually execute Jython in JythonServer are all
subclasses of JythonServerThread. This change sets the thread context
class loader (TCCL) to the Jython bundle loader when a new Jython thread
is created, which is for each command.

This should have no effect on any code that does not explicitly use the
TCCL.

Change-Id: Ieb905098209f4dd20df7e4a9334107707fde4233