Improve logging of varargs methods
The logging didn't work very well because varargs are passed as an array
which the logger also accepts to fill the message placeholders. So with
one placeholder only the first arg would be logged which could be
confusing.

This adds Arrays#asList to the args before logging them, this still
offers the benefit of paramerized logging as toString will not be called
on the List unless it will be logged.

Change-Id: I7ce7effcf5db0f71cccb4187ffb652b9b435f2b7