DAQ-1251 Move client Spring initialisation before user and visit setup
This is needed to move from Corba to RMI. This code used to explicitly
make and start an Corba AdaptorFactory and connect it to the Finder. It
was then used to get hold of server objects e.g GdaMetadata needed to
setup the visit. After this was done that AdaptorFactory was thrown
away and a new one was created by Spring during the createObjectFactory
method. This method call is now moved to the first thing done by the
client after determining a server is available.

Change-Id: I983dcf08ded39b4f4a477c3db276549b641d314a