DAQ-1280 Add motor limits validation to PolygonRegionEditor
Validation happens both in the EditingSupport and in the LabelProvider.
The EditingSupport works in the same way as the validation in the other editors, but is only visible when the cell is being edited. Therefore,
when the value lies outside the scannable limits, the LabelProvider sets
the foreground colour to red.

Change-Id: I89c94b9274fa216c072036c4ea4bdbd621dae294