example-config: fix scannableClasses Jython module
Replace references to deprecated classes or interfaces in the
scannableClasses Jython module with non-deprecated classes or
interfaces.  Without this, a NameError exception is raised when
attempting to import the module.

Change the ScannableGaussianWidth class in the scannableClasses module
to use the width attribute of the ScannableGaussian instance rather
than getWidth and setWidth methods which do not exist.  Also change the
ScannableGaussianWidth class docstring to not supply an initial value
in the constructor in its example since the constructor takes no such
value.

Change-Id: I18cbd2706716da7a26c571891bc4cdd402629450