[DAQ-1259] remove unnecessary DiseminateEvent inner class
This inner class is completely unnecessary as it just wraps the bean
received on the topic.
The special case DisseminateEvent.STOP object is also unnecessary as it
is only added to the queue when by disconnect() is called, which also
sets isConnected to false. Therefore the test for this object in the
thread created by createDiseminateThread() is never reached as the
while(isConnected()) will already have failed by this point, causing the
thread to exit.

Change-Id: Ic9fb43f9246e5b30314d10b7bb10abf6404dfb21