I15_1-537 (9.7pre) Fix thread leak in blockingSubmit & thus mscan
Fix blockingSubmit in SubmitterImpl so that a thread isn't consumed
every time it is called. This caused the total number of threads in
use to increase by 1 every time an mscan was run. This caused activemq
to break such that the activemq database had to be purged and the server
control machine had to be rebooted every time.

Also add various trace messages to track how and when submit and
blockingSubmit are called and when they complete.

Change-Id: I20f8736eef5752e84301a16dd49d9b96acac08f8