Modify PixiumNXDetector not to write file message to console
The message: Scan data will write to : <scanFilename>
can be misleading, as the code here assumes that the file name
will be prefixed by the beamline name (which is not true on I12).

It is also unnecessary, as the Nexus file writer will write a
similar (but correct) message.

There is also code in this function that sends a ScanStartEvent
containing the same files name (which may therefore also be
incorrect).

Change-Id: I7aa560b5671e99332934b05290e877123e75cb55